+++ shared/src/commonMain/kotlin/com/lszczepanski/fuelconsumptionlog/domain/model/Car.kt
@@ -7,7 +7,6 @@ data class Car(
     val engineCapacityCm3: Int,
     val horsePower: Int,
     val registrationNumber: String,
-    val initialMileageKm: Int,
-    val mileageKm: Int,
+    val initialMileageKm: Double,
+    val mileageKm: Double,
 )
-
